First two states moved to new architecture
