+++ src/c2m_controller.js
@@ -3,9 +3,13 @@
  */
 
 
-import { c2m_View } from './views/c2m_View.js';
+//import { c2m_View } from './views/c2m_View.js';
 
 import { c2m_InitialisedView } from './views/c2m_InitialisedView.js';
+import { c2m_ChooseWordView } from './views/c2m_ChooseWordView.js';
+import { c2m_CheckHtmlView } from './views/c2m_CheckHtmlView.js';
+import { c2m_CheckModuleView} from './views/c2m_CheckModuleView.js';
+import { c2m_CompletedView} from './views/c2m_CompletedView.js';
 
 import { c2m_Model } from './models/c2m_Model.js';
 
@@ -13,7 +17,7 @@ import { c2m_Model } from './models/c2m_Model.js';
 // Define the states
 
 const c2m_Initialised = "c2m_Initialised";
-const c2m_ChooseWord = "c2m_ChoseWord";
+const c2m_ChooseWord = "c2m_ChooseWord";
 const c2m_CheckHtml = "c2m_CheckHtml";
 const c2m_CheckModule = "c2m_CheckModule";
 const c2m_Completed = "c2m_Completed";
@@ -29,7 +33,7 @@ export default class c2m_Controller {
 		// Mammoth and Canvas Module converters??
 		this.model = new c2m_Model();
 		// TODO: will eventually create many different views
-		this.view = new c2m_View(this.model, this);
+//		this.view = new c2m_View(this.model, this);
 
 		// render the current state
 		this.render();
@@ -55,10 +59,54 @@ export default class c2m_Controller {
 		this.render();
 	}
 
+	/**
+	 * Event handler for uploading a Word doc
+	 * Use the model's convertWordDoc method, modify stage to checkHtml
+	 * and render
+	 */
 
+	handleUpload(event) {
+		console.log("handle upload");
+		console.log(event);
+		this.model.convertWordDoc(event);
 
+		// at this stage this.model.converter.mammothResult is defined
+		console.log("-------------------");
+		//console.log(this.model.converter.mammothResult); 
 
+		// get ready to display results
+		this.model.stage = c2m_CheckHtml;
+		this.render();
+	}
 
+	/**
+	 * Handle a mammoth result becoming available
+	 * Should only happen for checkHtml
+	 */
 
+	handleMammothResult(event) {
+		console.log("XXXXXXXXX mammoth result available");
+		console.log(this.model.converter.mammothResult);
+		// TODO update the div with the results
+		// handle any error messages
 
+		// Show the converted html
+		// update div#c2m_html with the result html
+		let c2m_html = document.getElementById("c2m_html");
+		if (c2m_html) {
+			c2m_html.innerHTML = this.model.converter.mammothResult.value;
+		}
+
+		// Show the messages from mammoth
+		let c2m_messages = document.getElementById("c2m_messages");
+		if (c2m_messages) {
+			let messageHtml = this.generateMessageHtml(this.model.converter.mammothResult.messages);
+			c2m_messages.innerHTML = messageHtml;
+		}
+
+		// hide div.c2m-waiting-results
+		document.querySelector("div.c2m-waiting-results").style.display = "none";
+		// display div.c2m-received-results
+		document.querySelector("div.c2m-received-results").style.display = "block";
+	}
 }
